moved the class to another directory
